상품등록 수정(완), 상품 목록 cri 진행중

+++ DongbaekMarketProject/src/main/java/com/itwillbs/domain/ProductVO.java
@@ -3,7 +3,7 @@ package com.itwillbs.domain;
 
 public class ProductVO {
 	private int product_code; //상품 코드
-	private long store_code; //가게 코드
+	private int store_code; //가게 코드
 	private String price; //상품 가격
 	private String product_name; //상품 이름
 	private int viewcnt; //조회수
